Refactored template converter
